update robot.py drive algorithm

+++ robot.py
@@ -34,8 +34,8 @@ class Lovebird():
             self.rightPower.write(abs(right))
 
         elif self.name == "Tessle":
-            self.leftPower.write( ( (leftMotor/255) *180) - 90 )
-            self.rightPower.write(( (rightMotor/255) *180) - 90)
+            self.leftPower.write( ( int(leftMotor/255) *180) - 90 )
+            self.rightPower.write(( int(rightMotor/255) *180) - 90)
